Legend Engine - Assortment of changes
1. refactor tests for simplified reading
2. autogenerate schema protocol models
3. grammar composer support tablesubquery
