+++ legend-engine-xts-sql/legend-engine-xt-sql-grammar/src/main/java/org/finos/legend/engine/language/sql/grammar/to/SQLGrammarComposer.java
@@ -527,7 +527,7 @@ public class SQLGrammarComposer
             @Override
             public String visit(TableSubquery val)
             {
-                return null;
+                return "(" + visit(val.query) + ")";
             }
 
             @Override
