+++ legend-engine-xts-sql/legend-engine-xt-sql-protocol/src/main/resources/core_external_query_sql_schema_metamodel.protocol.json
@@ -1,0 +1,6 @@
+{
+  "purePackage": "meta::external::query::sql::schema::metamodel",
+  "javaPackage": "org.finos.legend.engine.protocol.sql.schema.metamodel",
+  "elementsToBeExcluded": [
+  ]
+}
