+++ legend-engine-xts-sql/legend-engine-xt-sql-pure-metamodel/src/main/resources/core_external_query_sql_metamodel/schema_metamodel.pure
@@ -1,0 +1,41 @@
+Class meta::external::query::sql::schema::metamodel::SchemaColumn
+{
+  <<equality.Key>> name: String[1];
+}
+
+Class meta::external::query::sql::schema::metamodel::PrimitiveSchemaColumn extends meta::external::query::sql::schema::metamodel::SchemaColumn
+{
+  <<equality.Key>> type: meta::external::query::sql::schema::metamodel::PrimitiveType[1];
+}
+
+Class meta::external::query::sql::schema::metamodel::EnumSchemaColumn extends meta::external::query::sql::schema::metamodel::SchemaColumn
+{
+  <<equality.Key>> type: String[1];
+}
+
+
+Class meta::external::query::sql::schema::metamodel::Schema
+{
+  <<equality.Key>> columns: meta::external::query::sql::schema::metamodel::SchemaColumn[*];
+  <<equality.Key>> enums: meta::external::query::sql::schema::metamodel::Enum[*];
+}
+
+Enum meta::external::query::sql::schema::metamodel::PrimitiveType
+{
+  Boolean,
+  StrictDate,
+  Number,
+  String,
+  LatestDate,
+  Float,
+  DateTime,
+  Date,
+  Integer,
+  Decimal
+}
+
+Class meta::external::query::sql::schema::metamodel::Enum
+{
+  <<equality.Key>> type: String[1];
+  <<equality.Key>> values: String[*];
+}
