+++ legend-engine-xts-sql/legend-engine-xt-sql-query/pom.xml
@@ -33,10 +33,6 @@
             <groupId>org.finos.legend.pure</groupId>
             <artifactId>legend-pure-m3-core</artifactId>
         </dependency>
-        <dependency>
-            <groupId>org.finos.legend.pure</groupId>
-            <artifactId>legend-pure-runtime-java-engine-compiled</artifactId>
-        </dependency>
         <!-- Pure -->
 
         <!-- ENGINE -->
